# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the HLL App API"
  version: "2.0.0"
  title: "Hyreslandslaget - HLL API"
host: "api.hyreslandslaget.se"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /estimates/:
    get:
      tags:
      - "Estimate"
      summary: "List estimates"
      description: ""
      operationId: "estimatesGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          examples:
            application/json:
              items:
              - id: "AS3423"
                name: "Name"
                color: "#ff0055"
                price: "800 kr"
                modified: "2017-01-12T15:46:58Z"
                period:
                  start: "2017-01-12T15:46:58Z"
                  finish: "2017-01-12T15:46:58Z"
                assemblies:
                - id: 123
                  equipment:
                    article:
                      number: "1234567890"
                      name: "The Dumper"
                      description: "Lorem imposum"
                      sraCode: "211313"
                    offer:
                      name: "\"Hyra\""
                      amount: "123.45"
                      price: "12345 kr / alldager"
                      unit: "st"
                    services:
                      article:
                        number: "1234567890"
                        name: "The Dumper"
                        description: "Lorem imposum"
                        sraCode: "211313"
                      offer:
                        name: "\"Hyra\""
                        amount: "123.45"
                        price: "12345 kr / alldager"
                        unit: "st"
                    order:
                      amount: 1234.55999999999994543031789362430572509765625
                      discount: 123
                      price: "1234567.89 kr"
                  accessories:
                  - article:
                      number: "1234567890"
                      name: "The Dumper"
                      description: "Lorem imposum"
                      sraCode: "211313"
                    offer:
                      name: "\"Hyra\""
                      amount: "123.45"
                      price: "12345 kr / alldager"
                      unit: "st"
                      preSelected: true
                    services:
                      article:
                        number: "1234567890"
                        name: "The Dumper"
                        description: "Lorem imposum"
                        sraCode: "211313"
                      offer:
                        name: "\"Hyra\""
                        amount: "123.45"
                        price: "12345 kr / alldager"
                        unit: "st"
                    order:
                      amount: 1234.55999999999994543031789362430572509765625
                      discount: 123
                      price: "1234567.89 kr"
                discounts:
                  classes:
                  - code: "23"
                    name: "Some name"
                    value: 10
                  categories:
                  - code: "23"
                    name: "Some name"
                    value: 10
                insurance:
                  price: "150 kr"
                delivery:
                  destination:
                    latitude: "57.56236"
                    longitude: "18.34526"
                  fare:
                    price: "150 kr"
                  status: "ESTIMATED"
                  message: "A description of the status"
          headers: {}
          schema:
            type: "object"
            properties:
              items:
                type: "array"
                items: {}
        403:
          description: "Forbidden"
          examples:
            application/json:
              message: "Resource not allowed"
              code: 403
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
        404:
          description: "Not Found"
          examples:
            application/json:
              message: "Resource not found"
              code: 404
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
      x-swagger-router-controller: "Estimate"
    post:
      tags:
      - "Estimate"
      summary: "Create estimate"
      description: ""
      operationId: "estimatesPOST"
      produces:
      - "application/json"
      parameters:
      - name: "template"
        in: "query"
        description: "The ID of the template to base the new estimate on"
        required: false
        type: "string"
        x-example: "AS3423"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/body"
      responses:
        200:
          description: "OK"
          examples:
            application/json:
              id: "AS3423"
              name: "Name"
              color: "#ff0055"
              price: "800 kr"
              modified: "2017-01-12T15:46:58Z"
              period:
                start: "2017-01-12T15:46:58Z"
                finish: "2017-01-12T15:46:58Z"
              assemblies:
              - id: 123
                equipment:
                  article:
                    number: "1234567890"
                    name: "The Dumper"
                    description: "Lorem imposum"
                    sraCode: "211313"
                  offer:
                    name: "\"Hyra\""
                    amount: "123.45"
                    price: "12345 kr / alldager"
                    unit: "st"
                  services:
                    article:
                      number: "1234567890"
                      name: "The Dumper"
                      description: "Lorem imposum"
                      sraCode: "211313"
                    offer:
                      name: "\"Hyra\""
                      amount: "123.45"
                      price: "12345 kr / alldager"
                      unit: "st"
                  order:
                    amount: 1234.55999999999994543031789362430572509765625
                    discount: 123
                    price: "1234567.89 kr"
                accessories:
                - article:
                    number: "1234567890"
                    name: "The Dumper"
                    description: "Lorem imposum"
                    sraCode: "211313"
                  offer:
                    name: "\"Hyra\""
                    amount: "123.45"
                    price: "12345 kr / alldager"
                    unit: "st"
                    preSelected: true
                  services:
                    article:
                      number: "1234567890"
                      name: "The Dumper"
                      description: "Lorem imposum"
                      sraCode: "211313"
                    offer:
                      name: "\"Hyra\""
                      amount: "123.45"
                      price: "12345 kr / alldager"
                      unit: "st"
                  order:
                    amount: 1234.55999999999994543031789362430572509765625
                    discount: 123
                    price: "1234567.89 kr"
              discounts:
                classes:
                - code: "23"
                  name: "Some name"
                  value: 10
                categories:
                - code: "23"
                  name: "Some name"
                  value: 10
              insurance:
                price: "150 kr"
              delivery:
                destination:
                  latitude: "57.56236"
                  longitude: "18.34526"
                fare:
                  price: "150 kr"
                status: "ESTIMATED"
                message: "A description of the status"
          headers: {}
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                description: "The unique identifier"
              name:
                type: "string"
                description: "The name of estimate"
              color:
                type: "string"
                description: "The color of the estimate"
              price:
                type: "string"
                description: "The total productPrice"
              modified:
                type: "string"
                description: "ISO 8601 The last modified date"
              period:
                $ref: "#/definitions/estimates_period"
              assemblies:
                type: "array"
                description: "The rental items"
                items: {}
              discounts:
                $ref: "#/definitions/inline_response_200_6_discounts"
              insurance:
                $ref: "#/definitions/estimates_insurance"
              delivery:
                $ref: "#/definitions/estimates_delivery"
        403:
          description: "Forbidden"
          examples:
            application/json:
              message: "Resource not allowed"
              code: 403
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
        404:
          description: "Not Found"
          examples:
            application/json:
              message: "Resource not found"
              code: 404
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
      x-swagger-router-controller: "Estimate"
  /estimates/{id}:
    get:
      tags:
      - "Estimate"
      summary: "Get estimate"
      description: ""
      operationId: "estimatesIdGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          examples:
            application/json:
              id: "AS3423"
              name: "Name"
              color: "#ff0055"
              price: "800 kr"
              modified: "2017-01-12T15:46:58Z"
              period:
                start: "2017-01-12T15:46:58Z"
                finish: "2017-01-12T15:46:58Z"
              assemblies:
              - id: 123
                equipment:
                  article:
                    number: "1234567890"
                    name: "The Dumper"
                    description: "Lorem imposum"
                    sraCode: "211313"
                  offer:
                    name: "\"Hyra\""
                    amount: "123.45"
                    price: "12345 kr / alldager"
                    unit: "st"
                  services:
                    article:
                      number: "1234567890"
                      name: "The Dumper"
                      description: "Lorem imposum"
                      sraCode: "211313"
                    offer:
                      name: "\"Hyra\""
                      amount: "123.45"
                      price: "12345 kr / alldager"
                      unit: "st"
                  order:
                    amount: 1234.55999999999994543031789362430572509765625
                    discount: 123
                    price: "1234567.89 kr"
                accessories:
                - article:
                    number: "1234567890"
                    name: "The Dumper"
                    description: "Lorem imposum"
                    sraCode: "211313"
                  offer:
                    name: "\"Hyra\""
                    amount: "123.45"
                    price: "12345 kr / alldager"
                    unit: "st"
                    preSelected: true
                  services:
                    article:
                      number: "1234567890"
                      name: "The Dumper"
                      description: "Lorem imposum"
                      sraCode: "211313"
                    offer:
                      name: "\"Hyra\""
                      amount: "123.45"
                      price: "12345 kr / alldager"
                      unit: "st"
                  order:
                    amount: 1234.55999999999994543031789362430572509765625
                    discount: 123
                    price: "1234567.89 kr"
              discounts:
                classes:
                - code: "23"
                  name: "Some name"
                  value: 10
                categories:
                - code: "23"
                  name: "Some name"
                  value: 10
              insurance:
                price: "150 kr"
              delivery:
                destination:
                  latitude: "57.56236"
                  longitude: "18.34526"
                fare:
                  price: "150 kr"
                status: "ESTIMATED"
                message: "A description of the status"
          headers: {}
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                description: "The unique identifier"
              name:
                type: "string"
                description: "The name of estimate"
              color:
                type: "string"
                description: "The color of the estimate"
              price:
                type: "string"
                description: "The total productPrice"
              modified:
                type: "string"
                description: "ISO 8601 The last modified date"
              period:
                $ref: "#/definitions/estimates_period"
              assemblies:
                type: "array"
                description: "The rental items"
                items: {}
              discounts:
                $ref: "#/definitions/inline_response_200_6_discounts"
              insurance:
                $ref: "#/definitions/estimates_insurance"
              delivery:
                $ref: "#/definitions/estimates_delivery"
        403:
          description: "Forbidden"
          examples:
            application/json:
              message: "Resource not allowed"
              code: 403
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
        404:
          description: "Not Found"
          examples:
            application/json:
              message: "Resource not found"
              code: 404
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
      x-swagger-router-controller: "Estimate"
    delete:
      tags:
      - "Estimate"
      summary: "Delete estimate"
      description: ""
      operationId: "estimatesIdDELETE"
      produces:
      - "application/json"
      parameters: []
      responses:
        204:
          description: "No Content"
          examples: {}
          headers: {}
        403:
          description: "Forbidden"
          examples:
            application/json:
              message: "Resource not allowed"
              code: 403
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
        404:
          description: "Not Found"
          examples:
            application/json:
              message: "Resource not found"
              code: 404
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
      x-swagger-router-controller: "Estimate"
  /estimates/{id}?type=pdf:
    get:
      tags:
      - "Estimate"
      summary: "Get estimate as PDF"
      description: ""
      operationId: "estimatesIdtypepdfGET"
      produces:
      - "application/pdf"
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          examples:
            application/pdf: ""
          headers: {}
        403:
          description: "Forbidden"
          examples:
            application/json:
              message: "Resource not allowed"
              code: 403
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
        404:
          description: "Not Found"
          examples:
            application/json:
              message: "Resource not found"
              code: 404
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
      x-swagger-router-controller: "Estimate"
  /estimates/{id}/characteristics:
    put:
      tags:
      - "Estimate"
      summary: "Update estimate characteristics"
      description: ""
      operationId: "estimatesIdCharacteristicsPUT"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Estimate Characteristic"
      responses:
        200:
          description: "OK"
          examples:
            application/json:
              id: "AS3423"
              name: "Name"
              color: "#ff0055"
              price: "800 kr"
              modified: "2017-01-12T15:46:58Z"
          headers: {}
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                description: "The unique identifier"
              name:
                type: "string"
                description: "The name of estimate"
              color:
                type: "string"
                description: "The color of the estimate"
              price:
                type: "string"
                description: "The total productPrice"
              modified:
                type: "string"
                description: "ISO 8601 The last modified date"
        403:
          description: "Forbidden"
          examples:
            application/json:
              message: "Resource not allowed"
              code: 403
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
        404:
          description: "Not Found"
          examples:
            application/json:
              message: "Resource not found"
              code: 404
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "number"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
      x-swagger-router-controller: "Estimate"
...
definitions:
...
  estimates_period:
    required:
    - "finish"
    - "start"
    properties:
      start:
        type: "string"
        description: "ISO 8601 Inform about start date"
      finish:
        type: "string"
        description: "ISO 8601 Inform about finish date"
    description: "The rental period"
...
  body:
    type: "object"
    properties:
      id:
        type: "string"
        description: "The unique identifier"
      name:
        type: "string"
        description: "The name of estimate"
      color:
        type: "string"
        description: "The color of the estimate"
      price:
        type: "string"
        description: "The total productPrice"
      modified:
        type: "string"
        description: "ISO 8601 The last modified date"
      period:
        $ref: "#/definitions/estimates_period"
      assemblies:
        type: "array"
        description: "The rental items"
      discounts:
        $ref: "#/definitions/estimates_discounts"
      insurance:
        $ref: "#/definitions/estimates_insurance"
      delivery:
        $ref: "#/definitions/estimates_delivery"
    example:
      id: "AS3423"
      name: "Name"
      color: "#ff0055"
      price: "800 kr"
      modified: "2017-01-12T15:46:58Z"
      period:
        start: "2017-01-12T15:46:58Z"
        finish: "2017-01-12T15:46:58Z"
      assemblies:
      - id: 123
        equipment:
          article:
            number: "1234567890"
            name: "The Dumper"
            description: "Lorem imposum"
            sraCode: "211313"
          offer:
            name: "\"Hyra\""
            amount: "123.45"
            price: "12345 kr / alldager"
            unit: "st"
          services:
            article:
              number: "1234567890"
              name: "The Dumper"
              description: "Lorem imposum"
              sraCode: "211313"
            offer:
              name: "\"Hyra\""
              amount: "123.45"
              price: "12345 kr / alldager"
              unit: "st"
          order:
            amount: 1234.55999999999994543031789362430572509765625
            discount: 123
            price: "1234567.89 kr"
        accessories:
        - article:
            number: "1234567890"
            name: "The Dumper"
            description: "Lorem imposum"
            sraCode: "211313"
          offer:
            name: "\"Hyra\""
            amount: "123.45"
            price: "12345 kr / alldager"
            unit: "st"
            preSelected: true
          services:
            article:
              number: "1234567890"
              name: "The Dumper"
              description: "Lorem imposum"
              sraCode: "211313"
            offer:
              name: "\"Hyra\""
              amount: "123.45"
              price: "12345 kr / alldager"
              unit: "st"
          order:
            amount: 1234.55999999999994543031789362430572509765625
            discount: 123
            price: "1234567.89 kr"
      discounts:
        classes:
        - code: "23"
          name: "Some name"
          value: 10
        categories:
        - code: "23"
          name: "Some name"
          value: 10
      insurance:
        price: "150 kr"
      delivery:
        destination:
          latitude: "57.56236"
          longitude: "18.34526"
        fare:
          price: "150 kr"
        status: "ESTIMATED"
        message: "A description of the status"
  inline_response_200_6_discounts:
    properties:
      classes:
        type: "array"
        items: {}
      categories:
        type: "array"
        items: {}
    description: "The rental active discounts"
...